Refactor circuit breaker operator. Split into multiple classes.
